split chain requests
